Extract the absence qualifier so three surfaces share one implementation
Rung two adds the same verdict rendering to kin trace and kin search. Copying
impact's helper twice would be this ticket's own defect arriving by the door the
fix left open: the thing being closed is two surfaces answering one question
differently, and three copies is three chances to drift.

The module takes the tool name because it is load-bearing rather than
decorative. kin trace builds a context pack, so it declares get_context_pack and
never trace_data_flow, which describes a different walk; kin search declares
semantic_search, which reads no edge class and is gated on language scope
instead. Naming the wrong one gates a command on evidence it never gathered.

Behaviour for kin impact is unchanged: the observation stays impact-specific and
only the rendering moved.

FIR-2524
