+++ crates/kin-cli/src/commands/mod.rs
@@ -1,6 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright 2026 Firelock, LLC
 
+pub mod absence_qualifier;
 pub mod admit;
 pub mod agent;
 pub mod approvals;
